% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bodysites.R
\name{get_bodysite_projects}
\alias{get_bodysite_projects}
\title{Get Body Site Projects}
\usage{
get_bodysite_projects(body_site_id)
}
\arguments{
\item{body_site_id}{Body site ID (integer)}
}
\value{
Vector containing project accessions for the body site
}
\description{
Get all project accessions for a specific body site.
}
\examples{
projects <- get_bodysite_projects(2)
print(projects)
}
