% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bodysites.R
\name{get_bodysite_samples}
\alias{get_bodysite_samples}
\title{Get Body Site Samples}
\usage{
get_bodysite_samples(body_site_id)
}
\arguments{
\item{body_site_id}{Body site ID (integer)}
}
\value{
Vector containing sample run IDs for the body site
}
\description{
Get all sample runs for a specific body site.
}
\examples{
samples <- get_bodysite_samples(2)
print(samples)
}
