% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{get_countries}
\alias{get_countries}
\title{Get Country List}
\usage{
get_countries()
}
\value{
Matrix containing country information with columns: countryId, countryName, continent
}
\description{
Get all country information, including country ID, country name, and continent.
}
\examples{
countries <- get_countries()
print(dim(countries))
}
