% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{get_country_projects}
\alias{get_country_projects}
\title{Get Country Projects}
\usage{
get_country_projects(country_id)
}
\arguments{
\item{country_id}{Country ID (integer)}
}
\value{
Vector containing project accessions for the country
}
\description{
Get all project accessions for a specific country.
}
\examples{
projects <- get_country_projects(1)
print(projects)
}
