% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{get_country_samples}
\alias{get_country_samples}
\title{Get Country Samples}
\usage{
get_country_samples(country_id)
}
\arguments{
\item{country_id}{Country ID (integer)}
}
\value{
Vector containing sample run IDs for the country
}
\description{
Get all sample runs for a specific country.
}
\examples{
samples <- get_country_samples(1)
print(samples)
}
