% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypes.R
\name{get_phenotype_projects}
\alias{get_phenotype_projects}
\title{Get Phenotype Projects}
\usage{
get_phenotype_projects(phenotype_id)
}
\arguments{
\item{phenotype_id}{Phenotype ID (integer)}
}
\value{
Vector containing project accessions for the phenotype
}
\description{
Get all project accessions for a specific phenotype.
}
\examples{
projects <- get_phenotype_projects(1)
print(projects)
}
