% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypes.R
\name{get_phenotype_samples}
\alias{get_phenotype_samples}
\title{Get Phenotype Samples}
\usage{
get_phenotype_samples(phenotype_id)
}
\arguments{
\item{phenotype_id}{Phenotype ID (integer)}
}
\value{
Vector containing sample run IDs for the phenotype
}
\description{
Get all sample runs for a specific phenotype.
}
\examples{
samples <- get_phenotype_samples(1)
print(samples)
}
