% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypes.R
\name{get_phenotypes}
\alias{get_phenotypes}
\title{Get Phenotype List}
\usage{
get_phenotypes()
}
\value{
Matrix containing phenotype information with columns: phenotypeId, phenotypeName
}
\description{
Get all phenotypes with their IDs and names, including phenotype ID, phenotype name.
}
\examples{
phenotypes <- get_phenotypes()
print(dim(phenotypes))
}
