% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_project_metadata}
\alias{get_project_metadata}
\title{Get Project Metadata}
\usage{
get_project_metadata(project_accessions)
}
\arguments{
\item{project_accessions}{Character vector of project accession IDs}
}
\value{
Data frame containing project metadata
}
\description{
Retrieve project metadata by Project Accession(s).
}
\examples{
metadata <- get_project_metadata(c("PRJDB13875", "PRJDB16847"))
print(metadata)
}
