% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{get_project_samples}
\alias{get_project_samples}
\title{Get Project Samples}
\usage{
get_project_samples(project_accession)
}
\arguments{
\item{project_accession}{Project Accession ID(s) - can be a single string or vector of strings}
}
\value{
Character vector containing sample run IDs for a single project or a list of vectors containing sample run IDs for multiple projects
}
\description{
Get all sample runs for specific project(s).
}
\examples{
samples <- get_project_samples("PRJDB13875")
print(length(samples))
samples <- get_project_samples(c("PRJDB13875", "PRJDB16847"))
print(length(samples))
}
