% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{get_project_stats}
\alias{get_project_stats}
\title{Get Project Statistics}
\usage{
get_project_stats(project_accession)
}
\arguments{
\item{project_accession}{Project Accession ID}
}
\value{
List containing project statistics
}
\description{
Get detailed statistics for a specific project.
}
\examples{
\dontrun{
stats <- get_project_stats("PRJDB13875")
print(stats)
}
}
