% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{get_projects}
\alias{get_projects}
\title{Get Project List}
\usage{
get_projects()
}
\value{
Character vector containing project accessions
}
\description{
Get all project accession list.
}
\examples{
projects <- get_projects()
print(length(projects))
}
