% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{get_samples}
\alias{get_samples}
\title{Get Sample List}
\usage{
get_samples()
}
\value{
Matrix containing sample information with columns: run, projectAccession
}
\description{
Get all sample run list.
}
\examples{
samples <- get_samples()
print(dim(samples))
}
