% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{get_taxonomy_children}
\alias{get_taxonomy_children}
\title{Get Taxonomy Children}
\usage{
get_taxonomy_children(db, taxon_id)
}
\arguments{
\item{db}{Database type ("gg" or "silva")}

\item{taxon_id}{Taxon ID (integer)}
}
\value{
List containing taxonomy children information
}
\description{
Get all child taxa for a specific taxon.
}
\examples{
children <- get_taxonomy_children("gg", 15)
print(children)
}
