% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{get_taxonomy_list}
\alias{get_taxonomy_list}
\title{Get Taxonomy List}
\usage{
get_taxonomy_list(db, rank = "phylum")
}
\arguments{
\item{db}{Database type ("gg" or "silva")}

\item{rank}{Taxonomic rank (optional)}
}
\value{
Matrix containing taxonomy information with columns: taxonId, taxonName, rank, parentId, fullName
}
\description{
Get all taxa list for a specific database.
}
\examples{
taxa <- get_taxonomy_list("gg", "phylum")
print(taxa)
}
