% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{get_taxonomy_parent}
\alias{get_taxonomy_parent}
\title{Get Taxonomy Parent}
\usage{
get_taxonomy_parent(db, taxon_id)
}
\arguments{
\item{db}{Database type ("gg" or "silva")}

\item{taxon_id}{Taxon ID (integer)}
}
\value{
List containing taxonomy parent information
}
\description{
Get parent taxon information for a specific taxon.
}
\examples{
parent <- get_taxonomy_parent("gg", 15)
print(parent)
}
