% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{get_taxonomy_projects}
\alias{get_taxonomy_projects}
\title{Get Taxonomy Projects}
\usage{
get_taxonomy_projects(db, taxon_id)
}
\arguments{
\item{db}{Database type ("gg" or "silva")}

\item{taxon_id}{Taxon ID (integer)}
}
\value{
Vector containing project accessions for the taxon
}
\description{
Get all project accessions for a specific taxon.
}
\examples{
projects <- get_taxonomy_projects("gg", 15)
print(projects)
}
