% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{get_taxonomy_samples}
\alias{get_taxonomy_samples}
\title{Get Taxonomy Samples}
\usage{
get_taxonomy_samples(db, taxon_id)
}
\arguments{
\item{db}{Database type ("gg" or "silva")}

\item{taxon_id}{Taxon ID (integer)}
}
\value{
Vector containing sample run IDs for the taxon
}
\description{
Get all sample runs for a specific taxon.
}
\examples{
samples <- get_taxonomy_samples("gg", 15)
print(samples)
}
