% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{make_request}
\alias{make_request}
\title{Make HTTP Request to PRIME API}
\usage{
make_request(endpoint, method = "GET", body = NULL, timeout = 30)
}
\arguments{
\item{endpoint}{API endpoint path}

\item{method}{HTTP method (GET, POST)}

\item{body}{Request body for POST requests}

\item{timeout}{Request timeout in seconds}
}
\value{
API response
}
\description{
Make HTTP Request to PRIME API
}
