% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{parse_response}
\alias{parse_response}
\title{Parse API Response}
\usage{
parse_response(response)
}
\arguments{
\item{response}{HTTP response object}
}
\value{
Parsed JSON response
}
\description{
Parse API Response
}
